airflow_proj/requirements.txt
pandas
apache-airflow
pandas_gbq
apache-airflow-providers-google
